+++ electron/eme/shell.js
